shoppingAgent/requirements.txt
streamlit>=1.35.0
anthropic>=0.28.0
python-dotenv
serpapi
langgraph
requests
beautifulsoup4

